==> notebooks/00_prepare_dataset.ipynb <==
# Prepare Dataset

## About
This notebook contains the code to
1. download the ABO dataset 
2. Clean the dataset to extract title/product type
3. export dataset as HuggingFace compatible dataset

## Dataset

This notebook uses the [Amazon Berkeley Objects (ABO) Dataset](https://amazon-berkeley-objects.s3.amazonaws.com/index.html) . 

The dataset was created in partnership with Amazon and UC Berklely .

For 147,702 it contains product metadata , images and 3D models. 

%%bash 
cd ../artifacts/dataset_raw/amazon/
wget https://amazon-berkeley-objects.s3.amazonaws.com/archives/abo-listings.tar
tar -xvf abo-listings.tar

#!wget http://deepyeti.ucsd.edu/jianmo/amazon/metaFiles/All_Amazon_Meta.json.gz
#!wget https://amazon-berkeley-objects.s3.amazonaws.com/archives/abo-listings.tar

## Imports

import pathlib
import sklearn
import datasets
import pandas as pd
import sklearn.preprocessing
import sklearn.model_selection
import glob
import functools

!pwd

## Process Dataset

!ls ../artifacts/dataset_raw/amazon/listings/metadata

dataset_path_raw = "../artifacts/dataset_raw/amazon/listings/metadata"

glob.glob(f'{dataset_path_raw}/*.json.gz')

load all 16 files

df_raw = pd.concat(map(functools.partial(pd.read_json, lines=True ), 
                    glob.glob(f'{dataset_path_raw}/*.json.gz') )) 

df_raw.head()

len(df_raw)

sample record

df_raw.iloc[0].to_dict()

for this project, we only need `item_name` and `brand`.    
We can assume and take the first value for the fields

def parse_property(property_record:dict,property_name:str):
    try:
        r = property_record[property_name][0]
        if property_name =="node":
            return r['node_name']
        else:
            return r['value']
    except Exception as e:
        return None
    
def cleanup_record(raw_record:dict):
    
    
    record= {
        'brand': parse_property(raw_record,'brand')
        ,'item_id': raw_record['item_id']
        ,'item_name': parse_property(raw_record,'item_name')
        ,'product_type': parse_property(raw_record,'product_type')
        ,'node': parse_property(raw_record, 'node')
        , 'main_image_id': raw_record['main_image_id']
        ,'product_description': raw_record['product_description']

        
    }
    
    return pd.Series(record)

df = df_raw.apply(cleanup_record,axis=1)

df.head()

df.columns

df['product_type'].value_counts()

There are some product types that don't occur frequently.       
We should limit our training data to include at least 50+ product types

min_product_count = 500

compute top product types

top_products =  df['product_type'].value_counts().loc[lambda x: x>min_product_count].index.tolist()

len(df['product_type'].value_counts() ) , len (top_products)

df_all = df [ df['product_type'].isin(top_products) ].copy()


len(df_all)

`text` and `label` are the columns that are needed by Hugging Face Transformer package

Item title is the text. 
Product Type is the label we are predicting

df_all['label_name'] = df_all['product_type']
df_all['text'] = df_all['item_name']

encode the product type to a numeric label

label_encoder = sklearn.preprocessing.LabelEncoder()

label_encoder.fit(df_all['label_name'])

df_all['label'] = label_encoder.transform(df_all['label_name'])


Allocate 60% for training , 20% validation and 20% for training

df_train, df_test = sklearn.model_selection.train_test_split(df_all, train_size=.6, stratify= df_all['label'] )


df_test, df_val = sklearn.model_selection.train_test_split(df_test, test_size=.5, stratify= df_test['label'] )


print  ( 
{
    'train': len(df_train)
    ,'test': len(df_test)
    ,'val': len(df_val)
}

)

## Create Hugging Face Dataset

In order to later feed our model to HF transformers package, we need either Pytorch Dataloader or use HF [datasets](https://github.com/huggingface/datasets).

`Datasets` can easily be used by TF/ Pytorch


dataset_features = datasets.Features(
    {'text': datasets.Value('string')
     , 'item_name': datasets.Value('string')
     , 'label': datasets.ClassLabel(names=list ( label_encoder.classes_ ))
     , 'brand': datasets.Value('string')
     , 'item_id': datasets.Value('string')
     , 'main_image_id': datasets.Value('string')
    , 'node': datasets.Value('string')

    }

)

dataset_features.keys()

create dataset dictionary with all the subsets

interested_columns = dataset_features.keys()

dataset_train = datasets.Dataset.from_pandas(df_train[interested_columns],features=dataset_features)
dataset_test = datasets.Dataset.from_pandas(df_test[interested_columns],features=dataset_features)
dataset_validation = datasets.Dataset.from_pandas(df_test[interested_columns],features=dataset_features)

dataset_all = datasets.DatasetDict({
    'train': dataset_train,
    'test': dataset_test,
    'valid': dataset_validation }
)

dataset_all

dataset_all['train'][0]

all_classes = dataset_all['train'].features['label'].names_file
all_classes

## Persist Changes

save the dataset and load it 

dataset_path = '../artifacts/dataset_processed/'

dataset_all.save_to_disk(dataset_path)

datasets.load_from_disk(dataset_path)

# References

[Amazon Object Dataset](https://amazon-berkeley-objects.s3.amazonaws.com/index.html)       
[Hugging Face Tutorial on Custom Dataset](https://github.com/huggingface/notebooks/blob/master/transformers_doc/custom_datasets.ipynb)
