==> notebooks/02_timing.ipynb <==
# Inference Review : different models

## About

- Look at timing information of different bert models

## imports

import pathlib
import sklearn
import datasets
import pandas as pd
import torch

import numpy as np
import transformers
import os
import json
from ts.utils.util  import map_class_to_label
from tqdm import tqdm, trange
import time
import torchviz
import torch.nn as nn


query = "men shoes"

## Device Specs

print ( torch.__config__.parallel_info())

## Model: bert-large-uncased

base_model = "bert-large-uncased"

model = transformers.AutoModelForSequenceClassification.from_pretrained(base_model)

tokenizer = transformers.AutoTokenizer.from_pretrained(
                base_model
            )

res = tokenizer.encode_plus(query, return_tensors="pt",  padding="max_length", truncation=True)

res

%%timeit -r 3 -n 5

model_res = model(**res)
model_res

%%timeit -r 3 -n 5
if torch.cuda.is_available():
    model_cuda = model.cuda()
    model_cuda(res['input_ids'].cuda(),res['attention_mask'].cuda())

## Model: bert-base-uncased

base_model = "bert-base-uncased"

model = transformers.AutoModelForSequenceClassification.from_pretrained(base_model)

tokenizer = transformers.AutoTokenizer.from_pretrained(
                base_model
            )

res = tokenizer.encode_plus(query, return_tensors="pt",  padding="max_length", truncation=True)

%%timeit -r 3 -n 5

model_res = model(**res)
model_res

%%timeit -r 3 -n 5
if torch.cuda.is_available():
    model_cuda = model.cuda()
    model_cuda(res['input_ids'].cuda(),res['attention_mask'].cuda())

## Model: distilbert-base-uncased

base_model = "distilbert-base-uncased"

model = transformers.AutoModelForSequenceClassification.from_pretrained(base_model)

tokenizer = transformers.AutoTokenizer.from_pretrained(
                base_model
            )

res = tokenizer.encode_plus(query, return_tensors="pt",  padding="max_length", truncation=True)

%%timeit -r 3 -n 5

model_res = model(**res)
model_res

%%timeit -r 3 -n 5
if torch.cuda.is_available():
    model_cuda = model.cuda()
    model_cuda(res['input_ids'].cuda(),res['attention_mask'].cuda())
