==> notebooks/03_optimizing_model.ipynb <==
# Optimizing Model

## About

- Sample code to load artifacts / run inference
- Quantizing model
- Converting the Bert model with torch script

## Import

import pathlib
import sklearn
import datasets
import pandas as pd
import torch

import numpy as np
import transformers
import os
import json
from ts.utils.util  import map_class_to_label
from tqdm import tqdm, trange
import time
import utils

## Load model/tokenizer artifact

base_model= "distilbert-base-uncased"
model_dir =f'../artifacts/model/{base_model}'

load model

model = transformers.AutoModelForSequenceClassification.from_pretrained(model_dir, return_dict=False)

a utiliy method provided by torchserver requires labels to also be strings

id2label_str = {str(key): value for key, value in model.config.id2label.items()}
id2label_str

load tokenizer

tokenizer = transformers.AutoTokenizer.from_pretrained(
                model_dir
            )

use gpu if available

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')


## Load dataset

dataset_path = '../artifacts/dataset_processed'

raw_datasets = datasets.load_from_disk(dataset_path)

columns = set(raw_datasets['test'].column_names ) - set(['text','label'])
columns

set(raw_datasets.column_names )

tokenize the entire dataset

def tokenize_function(examples):
    return tokenizer(examples["text"], padding="max_length", truncation=True)

tokenized_datasets = raw_datasets.map(tokenize_function, batched=True, remove_columns=columns)

tokenized_datasets['test'][0].keys()

# size of input length of model

len ( tokenized_datasets['test'][0]['input_ids'] )

create a subset of the test dataset.
feel free to use the full dataset if running on GPU

subset = tokenized_datasets["test"].num_rows
subset = 20

test_dataset = tokenized_datasets["test"].shuffle(42).select(range(subset)) 
test_dataset.set_format(type='torch' )

test_dataset

## Predicting on Test dataset

utils.prediction_batch(model,test_dataset,device='cpu' )

## Optimization: Quantizing model

Pytorch supports three types of Quantization:
1. Dynamic Qunatization
2. Static Quantization
3. Qunatization Aware Training 

In this notebook, we look at Dyanmic Quantization.

**Dynamic Qunatization**, quantizes the weights . The activations are quantized on the fly.

Currently Pytorch doesn't support dynamic quantization on GPU 

model.to('cpu')

In the below code, we quantize all the `Linear` layers to int8 from float32.

quantized_model_int8 = torch.quantization.quantize_dynamic(
    model, {torch.nn.Linear}, dtype=torch.qint8
)
quantized_model_int8


How much file storage did we gain ? 

def print_size_of_model(model):
    torch.save(model.state_dict(), "temp.p")
    print('Size (MB):', os.path.getsize("temp.p")/1e6)
    os.remove('temp.p')

print_size_of_model(model)
print_size_of_model(quantized_model_int8)

The quantized model reduced the file storage by half, which also translates to less memory and faster execution .

### Benchmark Quantization

Lets measure how much speedup and possible loss in accuracy , we gained from quantization

test_dataset.select([0])

def time_model_evaluation(model, dataset=test_dataset,device:str='cpu'):

    if device =="cuda" and not torch.cuda.is_available():
        return 

    eval_start_time = time.time()
    
    result = utils.prediction_batch(model , dataset, device=device )
    
    eval_end_time = time.time()
    eval_duration_time_ms =  (eval_end_time - eval_start_time) * 1_000
    print(result)
    print("Evaluate total time (ms): {0:.1f}".format(eval_duration_time_ms))
    
def time_model_evaluation_single(model,dataset=test_dataset, device:str='cpu'):

    if device =="cuda" and not torch.cuda.is_available():
        return 
    
    model = model.to(device)

    tokens_tensor = dataset['input_ids'][0:1]
    masks_tensors = dataset['attention_mask'][0:1]


    tokens_tensor = tokens_tensor.to(device)
    masks_tensors = masks_tensors.to(device)
    

    eval_start_time = time.time()
    
    
    res = model(tokens_tensor, masks_tensors)
    
    eval_end_time = time.time()
    eval_duration_time_ms =  (eval_end_time - eval_start_time) * 1_000

    #print("Evaluate total time (ms): {0:.1f}".format(eval_duration_time_ms))

Evaluate the original FP32 BERT model on gpu

time_model_evaluation(model, device='cuda')

%%timeit -r 3 -n 5
time_model_evaluation_single(model, device='cuda')

Evaluate the original FP32 BERT model on cpu

time_model_evaluation(model, device='cpu')

%%timeit -r 3 -n 5
time_model_evaluation_single(model, device='cpu')

Evaluate the qunatized int8 BERT model

time_model_evaluation(quantized_model_int8, device='cpu')

%%timeit -r 5 -n 10
time_model_evaluation_single(quantized_model_int8, device='cpu')

dynamtic int8 qunatization is not supported on GPU

# time_model_evaluation(quantized_model_int8, device='cuda')
"""
Could not run 'quantized::linear_dynamic' with arguments from the 'CUDA' backend. 
This could be because the operator doesn't exist for this backend, or was omitted during 
the selective/custom build process (if using custom build). 
If you are a Facebook employee using PyTorch on mobile, 
please visit https://fburl.com/ptmfixes for possible resolutions. 
'quantized::linear_dynamic' is only available for these backends: [CPU, BackendSelect, Named, 
ADInplaceOrView, AutogradOther, AutogradCPU, AutogradCUDA, AutogradXLA, UNKNOWN_TENSOR_TYPE_ID, 
AutogradMLC, Tracer, Autocast, Batched, VmapMode].
"""

| Model | device | Accuracy | Time |
| --- | --- | --- |--- |
| original fp32 | cpu | .918 | 116 |
| original fp32 | gpu | .918 | 17 |
| quantized int8 | cpu | .919 | 83 |













the above results , were measured in the below config.  
You might need to reduce the OpenMP threads, if introducing parralelism in your application

print ( torch.__config__.parallel_info())

## Optimization: TorchScript

TorchScript is a way to create serializable and optimizable models from PyTorch code

load model as torchscript

According to [Hugging Face docs](https://huggingface.co/transformers/torchscript.html), the `torchscript` flag

>This flag is necessary because most of the language models in this repository have tied weights between their Embedding layer and their Decoding layer. 
>
>TorchScript does not allow the export of models that have tied weights, it is therefore necessary to untie the weights beforehand.
>
>This implies that models instantiated with the torchscript flag have their Embedding layer and Decoding layer separate, which means that they should not be trained down the line. Training would de-synchronize the two layers, leading to unexpected results.
>
>This is not the case for models that do not have a Language Model head, as those do not have tied weights. 
These models can be safely exported without the torchscript flag.

script_model = transformers.AutoModelForSequenceClassification.from_pretrained(model_dir, torchscript=True
                                                                               , return_dict=False)
script_tokenizer = transformers.AutoTokenizer.from_pretrained(model_dir , torchscript=True)

create a dummy input to pass to our model

text = "men shoes"
res = script_tokenizer.encode_plus(text, return_tensors="pt", padding="max_length",truncation=True)

tokens_tensor = res['input_ids']
masks_tensors = res['attention_mask']

dummy_input = [tokens_tensor, masks_tensors]

dummy_input

Creating the trace by passing a dummy input

script_model = script_model.to(device)

tokens_tensor = tokens_tensor.to(device)
masks_tensors = masks_tensors.to(device)

traced_model = torch.jit.trace(script_model, [tokens_tensor, masks_tensors])

Creating the trace of quantized model by passing a dummy input

script_model = script_model.to('cpu')
quantized_model_int8 = torch.quantization.quantize_dynamic(
    script_model, {torch.nn.Linear}, dtype=torch.qint8
)


quantized_model_int8 = quantized_model_int8.to('cpu')
dummy_input = (tokens_tensor.to('cpu'), masks_tensors.to('cpu') )
quantized_model_int8 = quantized_model_int8.to('cpu')
traced_model_int8 = torch.jit.trace(quantized_model_int8, dummy_input )  

lets compute how long it takes to predict with traced model

time_model_evaluation(traced_model, device='cpu')
#prediction_batch(traced_model,test_dataset )

%%timeit -r 3 -n 5
time_model_evaluation_single(traced_model, device='cpu')

time_model_evaluation(traced_model, device='cuda')

%%timeit -r 3 -n 5
time_model_evaluation_single(traced_model, device='cuda')

%%timeit -r 3 -n 5
time_model_evaluation_single(traced_model_int8, device='cpu')

utils.prediction( model=traced_model_int8
                 , tokens_tensor=tokens_tensor
                 , masks_tensors=masks_tensors 
                 , id2label_str=id2label_str)

| Model | device | Accuracy | Time |
| --- | --- | --- |--- |
| original fp32 | cpu | .918 | 130 |
| original fp32 (trace) | cpu | .918 | 124 |
| original fp32 | gpu | .918 | 22 |
| original fp32 (trace) | gpu | .918 | 18 |
| quantized int8 | cpu | .919 | 104 |
| quantized int8 (trace) | cpu | .919 | 89.5 |

Save / load  traced model

model_dir_trace = f'{model_dir}__trace'

!mkdir -p {model_dir_trace}

torch.jit.save(traced_model, f"{model_dir_trace}/traced_model.pt")

loaded_model = torch.jit.load( f"{model_dir_trace}/traced_model.pt")
loaded_model.eval()

## Save Artifacts

For the next section, let us save the class label

with open(f'{model_dir}/index_to_name.json','w') as f:
    json.dump(id2label_str,f)

with open(f'{model_dir_trace}/index_to_name.json','w') as f:
    json.dump(id2label_str,f)

in a separate json config, let use store things like 
- was lower case used by the tokenizer
- max length of tokenizer
- num labels
- is the model a jit

tokenizer.model_max_length

tokenizer.max_len_single_sentence

tokenizer.save_pretrained(model_dir_trace)

setup_config = {
 "model_name":"pt-original",
 "do_lower_case": tokenizer.do_lower_case,
 "num_labels":len(id2label_str),
 "save_mode":"original",
 "max_length":tokenizer.model_max_length,
 "captum_explanation": True,
 "base_model": base_model,
 "top_k": 5   

}

setup_config



with open(f'{model_dir}/setup_config.json','w') as f:
    json.dump(setup_config,f)

setup_config_trace = {**setup_config}
setup_config_trace['model_name'] = "pt-jit"
setup_config_trace['captum_explanation'] = False
setup_config_trace['save_mode'] = 'jit'
setup_config_trace

with open(f'{model_dir_trace}/setup_config.json','w') as f:
    json.dump(setup_config_trace,f)
